requirements.txt
smolagents
gradio
transformers
requests
markdownify
duckduckgo_search
pillow
pytz
pyyaml
